requirements.txt
numpy==2.1.3
pandas==2.2.3
binance-connector==3.6.0
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.2
